RAG/FromDocument/requirements.txt
streamlit
langchain
langchain-google-genai
chromadb
python-dotenv
unstructured
markdown
python-magic
nltk
nest-asyncio